renamed classes and use Bhepop2Enrichment as the main one
