+++ bhepop2/uniform_enrich.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 
-# TODO : refactor this function in a class similar to MaxEntropyEnrichment
+# TODO : refactor this function in a class similar to Bhepop2Enrichment
 def uniform_enrich(population, distributions, abs_minimum, relative_maximum, seed):
     """
